Refactored todoform into a modal btn
