+++ src/App.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import Routes from './utils/Routes';
+import "semantic-ui-css/semantic.min.css";
 import ReactGA from "react-ga"; 
 
 function initializeAnalytics (){
