+++ src/components/dashboard/Dashboard.js
@@ -2,6 +2,7 @@ import React from "react";
 import TodoList from "./todos/TodoList.js";
 import TodoForm from "./todo-form/TodoForm.js";
 import { Button } from "semantic-ui-react";
+import AddTodoBtn from './add-todo/AddTodoBtn.js';
 
 import "../../SASS/Dashboard.scss";
 
@@ -18,7 +19,7 @@ const Dashboard = props => {
         Logout
       </Button>
       <TodoList />
-      <TodoForm />
+      <AddTodoBtn />
     </>
   );
 };
