+++ src/components/dashboard/add-todo/AddTodoBtn.js
@@ -1,29 +1,46 @@
 import React from "react";
-import TodoForm from "./TodoForm";
 
-import {Button, Modal, Icon} from "semantic-ui-react";
+import { Button, Modal, Dropdown, Icon } from 'semantic-ui-react';
 
+import AssignTime from './AssignTime.js';
 
-const ModalButton = () => {
-    return (
-      <Button className="ui primary button">
+  const AddTodoBtn = () => (
+    <Button className="ui primary button">
           <Icon aria-hidden="true" className="add" />
-  
+
         </Button>
+  )
+
+  const AddTodoModal = () => {
+
+    const options = [
+      {icon: 'member', text: 'Mom' , value: 'Mom'},
+      {icon: 'member', text: 'Dad' , value: 'Dad'},
+      {icon: 'member', text: 'Daughter' , value: 'Daughter'},
+      {icon: 'member', text: 'Son' , value: 'Son'},
+    
+    ]
+
+    return (
+      <Modal centered trigger={AddTodoBtn()}>
+        <Modal.Header> Task Details </Modal.Header>
+        <Modal.Content>
+        <Modal.Description> Who would you like to assign this task? </Modal.Description>
+        <Dropdown
+            className= 'button icon'
+            floating
+            options={options}
+            trigger={<React.Fragment />} 
+              />
+              <AssignTime></AssignTime>
+
+        </Modal.Content>
+      </Modal>
     )
   }
 
-  const AddTodoBtn = () => (
-    <Modal trigger={ModalButton()}>
-      <Modal.Header> Task Details </Modal.Header>
-       <TodoForm></TodoForm>
-        </Modal>
 
-
-  )
-
-
-  export default AddTodoBtn;
+  export default AddTodoModal;
