+++ src/components/household/InviteConfirm.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { Loader, Dimmer } from 'semantic-ui-react';
+
+export const InviteConfirm = () => {
+
+    return (
+    <div>
+        
+    </div>
+    )
+}
